\name{finalresults}
\alias{finalresults}
\title{finalresults}
\usage{
finalresults()
}
\description{
This function use the optimal value of two parameters in Seurat (nFeatures, resolution) to output results when using a reference scRNA-seq data to identify the cell type of unmark data.
}
\examples{
finalresults()
}
